requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 科学计算和数据处理
numpy>=1.21.0
pandas>=1.3.0

# 可视化
matplotlib>=3.4.0
tqdm>=4.62.0

# 图像处理
Pillow>=8.2.0
opencv-python>=4.5.0

# 工具库
scikit-learn>=0.24.0
tensorboard>=2.8.0

# 其他依赖
pyyaml>=5.4.0 